+++ src/OFS/DTMLDocument.py
@@ -43,6 +43,7 @@ class DTMLDocument(PropertyManager, DTMLMethod):
     """
     meta_type = 'DTML Document'
     zmi_icon = 'far fa-file-alt'
+    _locked_error_text = 'This document has been locked.'
 
     manage_options = DTMLMethod.manage_options
 
@@ -51,29 +52,6 @@ class DTMLDocument(PropertyManager, DTMLMethod):
         (perms[0] == change_dtml_methods) and
         (change_dtml_documents, perms[1]) or perms
         for perms in DTMLMethod.__ac_permissions__])
-
-    def manage_upload(self, file='', REQUEST=None):
-        """ Replace the contents of the document with the text in 'file'.
-        """
-        self._validateProxy(REQUEST)
-        if self.wl_isLocked():
-            raise ResourceLockedError('This document has been locked.')
-
-        if REQUEST and not file:
-            raise ValueError('No file specified')
-
-        if hasattr(file, 'read'):
-            file = file.read()
-        if PY3 and isinstance(file, binary_type):
-            file = file.decode('utf-8')
-        if PY2 and isinstance(file, text_type):
-            file = file.encode('utf-8')
-
-        self.munge(file)
-        self.ZCacheable_invalidate()
-        if REQUEST:
-            message = "Content uploaded."
-            return self.manage_main(self, REQUEST, manage_tabs_message=message)
 
     def __call__(self, client=None, REQUEST={}, RESPONSE=None, **kw):
         """Render the document with the given client object.